requirements.txt
altair==4.2.2
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bokeh==3.5.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
entrypoints==0.4
exceptiongroup==1.2.1
Faker==26.0.0
favicon==0.7.0
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
highlight-text==0.2
htbuilder==0.6.2
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.1    #matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.3.0
mplsoccer==1.2.4
numpy==2.0.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
prometheus_client==0.20.0
protobuf==5.27.2
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.36.0
streamlit-aggrid==1.0.5
streamlit-antd-components==0.3.2
streamlit-bokeh-events==0.1.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-lottie==0.0.5
streamlit-searchbox==0.1.13
streamlit-searchbox==0.1.13
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.30.0
watchdog==4.0.1
webdriver-manager==4.0.1
wsproto==1.2.0
xyzservices==2024.6.0
